Refactor framework to automate garbage collector disabling and reinstating
The framework provided functions to get and reset the state of the garbage
collector and the responsibility was on the tracer writer to use them in the
hook code. Now the garbage collector is automatically disabled before hook
execution and reinstated after hook execution. This removes a big mental burden
for the tracer writer as failing to disable the garbage collector can lead to
insidious bugs. Besides, its also difficult to remember to reinstate the garbage
collector at each return point in the hook code.

This commit also appropriately renames the global variables used by the
framework by prefixing them with `RDT_`.