Fix event_type for promise_expression_lookup
When promise expression is retrieved by substitute function, the tracer
records it with the event_type 1. This is indistinguishable from the
normal promise value lookup which is also recorded with event_type 1.
This commit changes the event_type for promise expression lookup to 3.
The analysis can now use this to distinguish between value lookup and
expression lookup.