Fix circular dependency issue by delinking SqlSerializer and State
SqlSerializer accesses and modifies the global state object. This is bad
design as this hides the changes to the state object and also makes the
serializer rely on the state object. This commit changes the API of the
SqlSerializer so that it takes the state object's properties as
arguments. This also fixes a circular dependency issue which prevents
code from compiling.