Refactor the tracing framework and remvove RDT library
RDT library contained mostly non functional or not useful code.
This commit removes it.

Furthermore, the API of the tracing framework has been revamped with
consistent names and better features. Tracers now receive a context
information which they can use for storing State/Configuration of
the tracing. The API also ensures that SEXP arguments are protected
before they are passed to probes.