Fix checking of NULL objects for dyntracing and destruction
The dyntracer_destroy_sexp function for cleaning up dyntracers checks
if dyntracer is null before calling the user provided function for
cleaning up the object. This makes the user level destroy function
simpler and removes chances of crashes arising from double cleanup of
dyntracers.

The dyntrace function now checks if the dyntracer is NULL and throws
an error instead of continuing. Passing a NULL dyntracer can happen if
a dyntracer that has been destroyed is passed to dyntrace function for
dyntracing. This is a mistake and the framework now prints a meaningful
error message to notify the user.