Add support for accessing byte compiler method `findOp`
`findOp` function tells the opcode of the byte compiled expression.
This is useful for disassembly of the expression. However, this
function is declared `static`, making it inaccessible outside the
scope of `eval.c`. This commit makes it accessible by removing the
`static` keyword and providing a declaration in `Rdyntrace.h`.