Add logging functionality, change `do_dyntrace` behavior
This commit adds functionality to log errors, warnings and info.
These are implemented as macros which print useful information such
as file name, function name and line number from where they are
called.
`dyntrace_log_error` jumps to top level after printing the error.
This lets the users jump out of tracing when an error is encountered
but keeps them in the existing R session for more interactions.

The behavior of `do_dyntrace` function has been modified.
`DYNTRACE_PROBE_END` is no longer called when the dyntracing
encounters errors, either because of R code or because of
the tracer.

The code has been reformatted using `clang-format` to ensure uniform
indentation and style.

Resolves #160